fix(#67): 디자인/스토어 상세 페이지로 변경
