+++ src/app/mypage/approve/page.tsx
@@ -46,11 +46,11 @@ const ApprovePage = () => {
         headerType="DYNAMIC"
         title="픽업 확정"
         headerClassname={'fixed bg-white '}
-        className={'bg-white items-start '}
+        className={'bg-white items-start pt-20 pb-3'}
         description={'곧 만날 케이크들이에요!'}
       />
 
-      <section className="mt-[120px] flex flex-col gap-y-[1.5rem]">
+      <section className="mt-[150px] flex flex-col gap-y-[1.5rem]">
         {/* 픽업 예정 */}
         <section className="px-[1.25rem]">
           <h2 className="title-l p-[0.5rem]">
