+++ src/app/search/layout.tsx
@@ -6,7 +6,6 @@ const SearchLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <div>
       {children}
-      <NavBar />
     </div>
   )
 }
