+++ src/assets/svgComponents/LocationIcon.tsx
@@ -1,0 +1,4 @@
+import * as React from "react";
+import type { SVGProps } from "react";
+const SvgLocationIcon = (props: SVGProps<SVGSVGElement>) => <svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" fill="none" viewBox="0 0 23 22" {...props}><g stroke="#3E3E3E" strokeLinecap="round" strokeLinejoin="round"><path d="M11.2 11.915a2.75 2.75 0 1 0 0-5.5 2.75 2.75 0 0 0 0 5.5" /><path d="M11.2 20.165c3.666-3.666 7.333-6.95 7.333-11a7.333 7.333 0 1 0-14.667 0c0 4.05 3.667 7.334 7.333 11" /></g></svg>;
+export default SvgLocationIcon;
