+++ src/assets/svgComponents/index.ts
@@ -8,16 +8,23 @@ export { default as BlueClipboardIcon } from './BlueClipboardIcon'
 export { default as BlueClockIcon } from './BlueClockIcon'
 export { default as BlueDropDownIcon } from './BlueDropDownIcon'
 export { default as BluePencilIcon } from './BluePencilIcon'
+export { default as BlueSendIcon } from './BlueSendIcon'
 export { default as CakeIcon } from './CakeIcon'
+export { default as CakeLogo } from './CakeLogo'
 export { default as CameraIcon } from './CameraIcon'
 export { default as CancelIcon } from './CancelIcon'
 export { default as ChevronLeftIcon } from './ChevronLeftIcon'
 export { default as ChevronRightIcon } from './ChevronRightIcon'
+export { default as CopyIcon } from './CopyIcon'
+export { default as CreditCardCheckIcon } from './CreditCardCheckIcon'
 export { default as DarkGrayPlusIcon } from './DarkGrayPlusIcon'
 export { default as DownLineIcon } from './DownLineIcon'
+export { default as DropUpIcon } from './DropUpIcon'
 export { default as DropDownIcon } from './DropDownIcon'
 export { default as FillDropDownIcon } from './FillDropDownIcon'
+export { default as FlyingMoneyIcon } from './FlyingMoneyIcon'
 export { default as GalleryIcon } from './GalleryIcon'
+export { default as GlitterGroupIcon } from './GlitterGroupIcon'
 export { default as Gray300ClipboardIcon } from './Gray300ClipboardIcon'
 export { default as Gray300HeartIcon } from './Gray300HeartIcon'
 export { default as Gray700HeartIcon } from './Gray700HeartIcon'
@@ -30,16 +37,26 @@ export { default as GrayImagePlusIcon } from './GrayImagePlusIcon'
 export { default as GrayPlusIcon } from './GrayPlusIcon'
 export { default as GrayRightArrowIcon } from './GrayRightArrowIcon'
 export { default as GrayUncheckCalendarIcon } from './GrayUncheckCalendarIcon'
+export { default as GrayHeartIcon } from './GrayHeartIcon'
+export { default as GrayHomeIcon } from './GrayHomeIcon'
+export { default as GrayMypageIcon } from './GrayMypageIcon'
+export { default as GrayOrderIcon } from './GrayOrderIcon'
 export { default as GraySearchIcon } from './GraySearchIcon'
+export { default as GrayStoreIcon } from './GrayStoreIcon'
 export { default as HeartIcon } from './HeartIcon'
 export { default as HeartIconEmpty } from './HeartIconEmpty'
 export { default as HeartIconFill } from './HeartIconFill'
 export { default as ImagePlusIcon } from './ImagePlusIcon'
+export { default as KakaoAppIcon } from './KakaoAppIcon'
 export { default as KakaoIcon } from './KakaoIcon'
 export { default as KeywordDeleteIcon } from './KeywordDeleteIcon'
 export { default as LeftArrowIcon } from './LeftArrowIcon'
+export { default as LocationIcon } from './LocationIcon'
 export { default as LogoutIcon } from './LogoutIcon'
 export { default as MarkerPinIcon } from './MarkerPinIcon'
+export { default as MemoSearchIcon } from './MemoSearchIcon'
+export { default as MemofileIcon } from './MemofileIcon'
+export { default as MessageIcon } from './MessageIcon'
 export { default as PlusIcon } from './PlusIcon'
 export { default as ProfileIcon } from './ProfileIcon'
 export { default as RedHeartIcon } from './RedHeartIcon'
@@ -49,6 +66,7 @@ export { default as RedSearchIcon } from './RedSearchIcon'
 export { default as SearchIconGray } from './SearchIconGray'
 export { default as SearchIconRed } from './SearchIconRed'
 export { default as SimpleCalendarIcon } from './SimpleCalendarIcon'
+export { default as TermsOfUseBlueCheck } from './TermsOfUseBlueCheck'
 export { default as WhiteCalendarIcon } from './WhiteCalendarIcon'
 export { default as WhiteCancelIcon } from './WhiteCancelIcon'
 export { default as WhiteCheckIcon } from './WhiteCheckIcon'
