+++ src/components/home/HomePage.tsx
@@ -4,27 +4,17 @@ import Header from '@/components/common/Header'
 import NavBar from '@/components/common/NavBar'
 import SearchInput from '@/components/home/SearchInput'
 import { RedSearchIcon } from '@/assets/svgComponents'
-import { useEffect, useState } from 'react'
+import { useEffect, useRef, useState } from 'react'
 import { AnimatePresence, motion } from 'framer-motion'
-import GuardianModal from '@/components/authentication/GuardianModal'
 import TotalSearchPage from '@/components/search/TotalSearchPage'
 import { useLoginStore } from '@/store/authStore'
 import SearchContent from '@/components/search/SearchContent'
 import WelcomeModal from '@/components/authentication/WelcomeModal'
 import useScrollDirection from '@/hooks/useScrollDirection'
 import Filter from '@/components/common/Filter'
-import StoreDetail from '@/components/search/StoreDetail'
-import DesignDetailContent from '@/components/search/DesignDetailContent'
 import useSearchResult from '@/hooks/useSearchResult'
-import OrderForm from '@/components/order/OrderForm'
 import { useSearchStore } from '@/store/searchStore'
-import { Drawer } from '@/components/ui/drawer'
-import OrderSubmissionSuccessModal from '@/components/order/OrderSubmissionSuccessModal'
 import Onboarding from '@/components/onboarding/Onboarding'
-import OrderExitConfirmModal from '@/components/mypage/OrderExitConfirmModal'
-import { useOrderStore } from '@/store/orderStore'
-import RequireLoginModal from '@/components/mypage/RequireLoginModal'
-import Cookies from 'js-cookie'
 
 const HomePage = () => {
   const isTotalSearchPageOpen = useSearchStore((state) => state.isTotalSearchPageOpen)
@@ -34,24 +24,15 @@ const HomePage = () => {
   const isGuardianModalOpen = useLoginStore((state) => state.isGuardianModalOpen)
 
   const {
-    isStoreDetailModalOpen,
-    isDesignDetailModalOpen,
     isFilterModalOpen,
-    isOrderFormOpen,
     setIsFilterModalOpen,
     selectedFilterType,
     setSelectedFilterType,
-    storeDetailMenu,
-    setStoreDetailMenu,
     totalCount,
     setSearchParams,
-    designDetail,
     renderFilterContent,
     keyword,
-    isOrderSubmissionSuccessModalOpen,
-    isLoginRequiredForOrderFormOpen,
     setState,
-    resetOrderForm,
     hasUserSelectedPrice,
   } = useSearchResult()
 
@@ -65,19 +46,6 @@ const HomePage = () => {
       return () => clearTimeout(timer) // cleanup
     }
   }, [isWelcomeModalOpen, setState])
-
-  // 3초 후 자동 닫힘 처리
-  useEffect(() => {
-    if (isOrderSubmissionSuccessModalOpen) {
-      const timer = setTimeout(() => {
-        setState({ isOrderSubmissionSuccessModalOpen: false })
-        //초기화
-        resetOrderForm()
-      }, 3000)
-
-      return () => clearTimeout(timer) // cleanup
-    }
-  }, [isOrderSubmissionSuccessModalOpen, setState])
 
   const scrollDirection = useScrollDirection()
 
@@ -98,37 +66,9 @@ const HomePage = () => {
     </AnimatePresence>
   ) : isTotalSearchPageOpen ? (
     <TotalSearchPage keyword={keyword} />
-  ) : isOrderFormOpen ? (
-    <OrderForm />
   ) : (
     <main className="bg-bg-300 relative min-h-screen">
-      <Drawer
-        open={isStoreDetailModalOpen || isDesignDetailModalOpen}
-        onOpenChange={(open) => {
-          if (!open) {
-            if (isStoreDetailModalOpen) {
-              setSearchParams({ isStoreDetailModalOpen: false })
-            }
-            if (isDesignDetailModalOpen) {
-              setSearchParams({ isDesignDetailModalOpen: false })
-            }
-          }
-        }}
-      >
-        {/* 로그인 안할 경우 주문서 대신 로그인 요구 모달 */}
-        {isLoginRequiredForOrderFormOpen && (
-          <RequireLoginModal
-            title={'주문하기'}
-            onClick={() => setState({ isLoginRequiredForOrderFormOpen: false })}
-            onCancelClick={() => setState({ isLoginRequiredForOrderFormOpen: false })}
-          />
-        )}
-
-        {/* 주문서 문의가 완료될 때 보이는 모달 */}
-        {isOrderSubmissionSuccessModalOpen && (
-          <OrderSubmissionSuccessModal onClick={() => setState({ isOrderSubmissionSuccessModalOpen: false })} />
-        )}
-
+      <div>
         {/* 회원가입 환영 모달 */}
         {isWelcomeModalOpen && (
           <AnimatePresence>
@@ -142,14 +82,6 @@ const HomePage = () => {
             <Filter setIsFilterModalOpen={setIsFilterModalOpen}>{renderFilterContent(selectedFilterType)}</Filter>
           )}
         </AnimatePresence>
-
-        {/* 가게 상세 페이지 모달 */}
-        {isStoreDetailModalOpen && (
-          <StoreDetail setStoreDetailMenu={setStoreDetailMenu} storeDetailMenu={storeDetailMenu} />
-        )}
-
-        {/* 디자인 상세 페이지 모달 */}
-        {isDesignDetailModalOpen && <DesignDetailContent designDetail={designDetail} />}
 
         <AnimatePresence>
           {isAtTop && (
@@ -187,10 +119,10 @@ const HomePage = () => {
           <SearchContent
             hasUserSelectedPrice={hasUserSelectedPrice}
             searchMenuClassname={
-              !isAtTop ? 'transition-all duration-300 fixed top-31' : 'transition-all duration-300 fixed top-60'
+              !isAtTop ? 'transition-all duration-300 fixed top-[127px]' : 'transition-all duration-300 fixed top-60'
             }
             FilterPanelClassname={
-              !isAtTop ? 'transition-all duration-300 fixed top-42' : 'transition-all duration-300 fixed top-71'
+              !isAtTop ? 'transition-all duration-300 fixed top-[171px]' : 'transition-all duration-300 fixed top-71'
             }
             SearchSummaryPanelClassname={!isAtTop ? 'mt-5' : 'mt-[9.375rem]'}
             isFilterModalOpen={isFilterModalOpen}
@@ -201,7 +133,7 @@ const HomePage = () => {
           />
         </div>
         <NavBar />
-      </Drawer>
+      </div>
     </main>
   )
 }
