+++ src/components/search/StoreProfile.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image'
-import { Gray700HeartIcon, HeartIcon, HeartIconFill } from '@/assets/svgComponents'
+import { Gray700HeartIcon, HeartIcon, HeartIconFill, LocationIcon } from '@/assets/svgComponents'
 import { DrawerTitle } from '@/components/ui/drawer'
 import { useHeartClick } from '@/hooks/useHeartClick'
 import { EventApi } from '@/api/eventAPI'
@@ -43,21 +43,9 @@ const StoreProfile = (props: Props) => {
   }
 
   return (
-    <div className="flex w-full flex-col items-center">
-      <section
-        className="flex w-full justify-end gap-x-[0.25rem] px-[1.438rem] pt-[1.25rem]"
-        onClick={(e) => {
-          handleHeartClick(storeId)
-        }}
-      >
-        {isLiked ? <HeartIconFill width={24} height={24} /> : <Gray700HeartIcon width={20} height={18} />}
-        <p className="caption-m text-gray-700">{likeCount}</p>
-      </section>
-      <DrawerTitle className="title-xl">{storeName}</DrawerTitle>
+    <div className="fixed top-28 pt-2 bg-white flex w-full flex-col items-center pb-[26.5px]">
       <div className="mt-[0.25rem] flex gap-x-[0.313rem]">
-        <div className="relative h-[1.375rem] w-[1.375rem]">
-          <Image src="/search/location.svg" alt="장소" fill className="object-cover"></Image>
-        </div>
+        <LocationIcon width={22} height={22} />
         <p className="body-m-m text-gray-700">{storeAddress}</p>
       </div>
