+++ src/components/search/StoreReview.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image'
 
 const StoreReview = () => {
   return (
-    <div className="mt-[3.313rem] mb-[3.313rem] flex h-full flex-col items-center justify-center gap-y-[0.938rem]">
+    <div className="flex flex-1 flex-col items-center justify-center min-h-[60vh] py-12 gap-y-[0.938rem]">
       <Image src={'/search/memo.svg'} width={28} height={32} alt={'메모'} style={{ width: 28, height: 32 }} />
       <div className="flex flex-col items-center justify-center">
         <div className="title-l text-gray-500">아직 리뷰는 준비중이에요!</div>
